// apps/mobile/src/api/generated/models/progressReportResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gota Vital API
 * API do app de nutrição e acompanhamento de saúde.
 * OpenAPI spec version: 0.1.0
 */
import type { BodyMeasurementResponseDto } from './bodyMeasurementResponseDto';
import type { DiaryEntryResponseDto } from './diaryEntryResponseDto';
import type { GoalTargetResponseDto } from './goalTargetResponseDto';

export interface ProgressReportResponseDto {
  generatedAt: string;
  weights: BodyMeasurementResponseDto[];
  goals: GoalTargetResponseDto[];
  diaryEntries: DiaryEntryResponseDto[];
}